V2.0.7 fix cpu compile warning

+++ libhesai/PtcClient/src/ptc_client.cc
@@ -95,7 +95,7 @@ bool PtcClient::IsValidRsp(u8Array_t &byteStreamIn) {
     byteStreamIn.erase(byteStreamIn.begin());
   }
   //输入数据长度小于头长度 没有数据
-  if (byteStreamIn.size() < ptc_parser_->GetPtcParserHeaderSize()) ret = false;
+  if (byteStreamIn.size() < (size_t)ptc_parser_->GetPtcParserHeaderSize()) ret = false;
 
   if (ret) {
     if(ptc_version_ == 1) {
@@ -117,7 +117,7 @@ void PtcClient::TcpFlushIn() {
     len = client_->Receive(u8Buf.data(), u8Buf.size(), MSG_DONTWAIT);
     if (len > 0) {
       std::cout << "TcpFlushIn, len" << len << std::endl;
-      for(int i = 0; i<u8Buf.size(); i++) {
+      for(size_t i = 0; i<u8Buf.size(); i++) {
         printf("%x ", u8Buf[i]);
       }
     }
@@ -134,7 +134,7 @@ int PtcClient::QueryCommand(u8Array_t &byteStreamIn,
   // TcpFlushIn();
 
   int nLen = client_->Send(encoded.data(), encoded.size());
-  if (nLen != encoded.size()) {
+  if (nLen != (int)encoded.size()) {
     // qDebug("%s: send failure, %d.", __func__, nLen);
     ret = -1;
   }
@@ -440,7 +440,7 @@ void PtcClient::CRCInit() {
 uint32_t PtcClient::CRCCalc(uint8_t *bytes, int len) {
   uint32_t i_crc = 0xffffffff;
   int i = 0;
-  for (i = 0; (size_t)i < len; i++)
+  for (i = 0; i < len; i++)
     i_crc = (i_crc << 8) ^ m_CRCTable[((i_crc >> 24) ^ bytes[i]) & 0xff];
   return i_crc;
 }
